Updated solutions.html
After reading the W3 validation errors for 'solution.html' I realised that some code had been put in after the initial validation that was then not used anyway. This was code that turned the locks red but that was found to not be suitable for the accessibility as seen by Lighthouse. This was easily rectified by taking out that unwanted code with id="contact_bolts" which was no longer needed.